requirements.txt
mysql-connector-python
requests
pandas
toolforge
python-dateutil